requirements.txt
absl-py==2.3.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.11.12
cffi==1.17.1
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
executing==2.2.1
Flask==3.1.2
flatbuffers==25.2.10
fonttools==4.59.2
gunicorn==23.0.0
idna==3.11
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml_dtypes==0.5.3
nest-asyncio==1.6.0
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
protobuf==4.25.8
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.2
pyngrok==7.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.1.0
requests==2.32.5
scipy==1.16.1
six==1.17.0
sounddevice==0.5.2
spotipy==2.25.2
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
urllib3==2.6.2
uv==0.8.15
wcwidth==0.2.13
Werkzeug==3.1.4
